+++ Changelog.md
@@ -17,6 +17,7 @@ Released on ?.
 
 - Fixed displaying sunrise and sunset when altitude is below 0 m
 - Fixed hovering over height profile to display track location
+- Fixed crash when location is requested but location services are disabled
 
 ## Version 1.15.0